+++ src/components/map/controls/swipe/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useCallback, useEffect } from 'react';
 
 import type BaseEvent from 'ol/events/Event';