+++ src/containers/hub/datasets-grid/constants.ts
@@ -1,0 +1,3 @@
+import type { SortingCriteria } from './types';
+
+export const SORTING = ['title', 'date'] satisfies SortingCriteria[];